All lepton-to-jet variables to the output nanoAOD Ntuple
We're unable to reproduce lepton-to-jet pT variables, because raw jet pT, raw correction factor and L1FastJet correction factors all different. Since we cannot get even the raw jet pT back, it indicates that the jet collection linked to the leptons are different from what is written to the Ntuple as ,,Jet'' collection.
Because the lepton-to-jet variables are computed anyways in order to evaluate the tth lepton MVA, there's very little computation overhead in adding the variables to the Ntuple. The only effect is that the output file size increases a little bit.
With this commit we can drop the nanoAOD-tools module that computes the lepton-to-jet variables. Sync will tell..