Add DeepCSV score of the associated jets to the leptons
Also, don't limit the b-tagging scores of the associated jets to be postive-semidefinite, because this restriction is necessary only in the evaluation of leptonMVA score (where the limit is already taken) and in synchronization, which can be done in post. By keeping the original b-tagging scores and setting the b-tagging scores to -1 for leptons that have no associated jet (instead of setting them to 0 in both cases), we gain information about wether the lepton has an associated jet or not. It doesn't affect the b-tagging cuts on the associated jets b/c the cuts always require the b-tagging scores to be less than some working point, not greater.